+++ CHANGES.rst
@@ -1,10 +1,12 @@
-threedi-qgis-plugin changelog
-=============================
+ThreeDiToolBox changelog
+========================
 
 
 0.16 (unreleased)
 -----------------
 - improve feedback to user in case of graph tool with v2_pumpstation_view
+
+- Close result selection window when the Escape key is pressed.
 
 - Fix leakage name.